Вынес реализацию удаленного репозитория в отдельный класс.
Теперь точно 1 ДЗ всё.
