+++ app/src/main/java/com/tashevv/poplibsgithub/App.kt
@@ -3,11 +3,12 @@ package com.tashevv.poplibsgithub
 import android.app.Application
 import android.content.Context
 import androidx.fragment.app.Fragment
-import com.tashevv.poplibsgithub.data.UsersRepoImpl
+import com.tashevv.poplibsgithub.data.UsersRepoLocalImpl
+import com.tashevv.poplibsgithub.data.UsersRepoRetrofitImpl
 import com.tashevv.poplibsgithub.domain.UsersRepo
 
 class App : Application() {
-    val usersRepo: UsersRepo by lazy { UsersRepoImpl() }
+    val usersRepo: UsersRepo by lazy { UsersRepoRetrofitImpl() }
 }
 
 val Context.app: App get() = applicationContext as App
