+++ app/src/main/java/com/tashevv/poplibsgithub/data/UsersRepoLocalImpl.kt
@@ -3,14 +3,10 @@ package com.tashevv.poplibsgithub.data
 import android.os.Handler
 import android.os.Looper
 import com.tashevv.poplibsgithub.domain.UserEntity
-import com.tashevv.poplibsgithub.domain.retrofit.UsersListRetrofitImpl
 import com.tashevv.poplibsgithub.domain.UsersRepo
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 
-class UsersRepoImpl : UsersRepo {
+class UsersRepoLocalImpl : UsersRepo {
 
     private val localData: List<UserEntity> = mutableListOf(
         UserEntity(
@@ -45,35 +41,9 @@ class UsersRepoImpl : UsersRepo {
         ),
     )
 
-    private var data: List<UserEntity> = localData
-
-    private val retrofitImpl = UsersListRetrofitImpl()
-
-    //    TODO remote data source
-    private fun remoteData() = retrofitImpl.getRetrofitImpl().getUsersList().enqueue(callback)
-
-    private val callback = object : Callback<List<UserEntity>> {
-
-        override fun onResponse(
-            call: Call<List<UserEntity>>,
-            response: Response<List<UserEntity>>
-        ) {
-            if (response.isSuccessful && response.body() != null) {
-                data = response.body()!!
-            } else Throwable(response.errorBody().toString())
-        }
-
-        override fun onFailure(call: Call<List<UserEntity>>, t: Throwable) {
-            data = localData
-        }
-
-    }
-
     override fun getUsers(onSuccess: (List<UserEntity>) -> Unit, onError: ((Throwable) -> Unit)?) {
-        remoteData()
-
         Handler(Looper.getMainLooper()).postDelayed({
-            onSuccess(data)
+            onSuccess(localData)
         }, 3_000L)
     }
 }
